+++ commercialNetwork/src/scss/index.scss
@@ -16,6 +16,7 @@
 @import "./merch.scss";
 @import "./scrollBar.scss";
 @import "./slider.scss";
+@import "./form.scss";
 
 body {
   overflow: hidden;
