+++ commercialNetwork/src/views/index.html
@@ -746,9 +746,10 @@
                 </div>
               </div>
             </div>
+
             <div class="popup__data">
               <div class="data__container">
-                <form action="" class="form" id="unique-form">
+                <form action="" class="form" id="form">
                   <div class="data__form">
                     <div class="form__column">
                       <div class="column__input">
@@ -760,12 +761,14 @@
                           <div data-error-message-city class="error-message"></div>
                         </div>
                       </div>
+
                       <div class="column__input">
                         <label class="input__title" for="job"> Должность на которую вы претендуете</label>
                         <input id="job" type="text" class="input__line" placeholder="Укажите должность" />
 
                         <div data-error-message-job class="error-message"></div>
                       </div>
+
                       <div class="column__input">
                         <label class="input__title" for="days">Дата рождения</label>
                         <div class="input__birthday">
@@ -781,6 +784,7 @@
                         </div>
                         <div class="error-message"></div>
                       </div>
+
                       <div class="column__input">
                         <label class="input__title" for="tel">Контактный телефон</label>
                         <input
@@ -793,17 +797,20 @@
                         <div data-error-message-tel class="error-message"></div>
                       </div>
                     </div>
+
                     <div class="form__column">
                       <div class="column__input">
                         <label class="input__title" for="email">Email адрес</label>
                         <input id="email" type="email" class="input__line" placeholder="Укажите ваш Email " />
                         <div data-error-message-email class="error-message"></div>
                       </div>
+
                       <div class="column__input">
                         <label class="input__title" for="name">Фамилия Имя Отчество</label>
-                        <input id="name" type="text" class="input__line" placeholder="Укажите ваше ФИО" />
-                        <div data-error-message-name class="error-message"></div>
+                        <input id="fio" type="text" class="input__line" placeholder="Укажите ваше ФИО" />
+                        <div data-error-message-fio class="error-message"></div>
                       </div>
+
                       <div class="column__input">
                         <label class="input__title" for="nationality">Гражданство</label>
                         <input
@@ -814,6 +821,7 @@
                         />
                         <div data-error-message-nationality class="error-message"></div>
                       </div>
+
                       <label class="input__button" for="file"
                         >Прикрепить резюме
                         <input
@@ -823,12 +831,18 @@
                           accept="application/pdf,application/vnd.ms-excel,.doc,.docx,.xml,application/msword,application/vnd.openxmlformats-officedocument.wordprocessingml.document"
                         />
                       </label>
-                      <div class="file-info">
-                        <span class="filename"></span>
+
+                      <div class="file-text">
+                        <div class="file-info">
+                          <span class="filename"></span>
+                          <span class="close-filename"> </span>
+                        </div>
+
                         <div data-error-message-file class="error-message"></div>
                       </div>
                     </div>
                   </div>
+
                   <div class="form__confirm">
                     <label class="confirm__container"
                       >Согласен с политикой обработки персональных данных
